% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoolNetPerturb.R
\name{attractorsLists2Dataframe}
\alias{attractorsLists2Dataframe}
\title{Convert a list of attractors to a data frame.}
\usage{
attractorsLists2Dataframe(attr.list, sep = "/")
}
\arguments{
\item{attr.list}{list of BoolNet attractor objects}

\item{sep}{string to join elements with length > 1, default "/"}
}
\value{
Dataframe, each column corresponds to a property of the attractor
}
\description{
Convert a list of BoolNet attractor objects to a data frame. Each property of eacn attr$attractors corresponds to a dataframe column. Columns are named attrName.propertyName, if the list has no names numbers will be used. If the propertie has elements with length > 1 it converts them to a string and joins them with sep.
}
\examples{
attractorsLists2Dataframe(attr.list)
}

