% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoolNetPerturb.R
\name{labelAttractors}
\alias{labelAttractors}
\title{Label attractor.}
\usage{
labelAttractors(attr, node.names, labels, rules, sep = "")
}
\arguments{
\item{attr}{Boolnet attractor object}

\item{node.names}{node names of the network}

\item{labels}{label that will be appended if the corresponding rule is TRUE}

\item{rules}{rules used for labelling the state. If the state satisfies the rule the corresponing label will be appended.  All the node names present in the rules must be in node.names}

\item{sep}{string to separate the labels when more than one can be applied to the state.}
}
\value{
List of strings corresponding to the label of the attractor, if an attractor has multiple states it will return a list of strings for that state.
}
\description{
Returns a list of labels for a the involved states in a BoolNet attractor object using a set of labelling rules, returns a list of labels.
}
\examples{
attr <- getAttractors(net$genes)
labelAttractors(attr, net$genes, labels, rules, sep='')
}
\seealso{
\code{\link{labelState}} which this function wraps
}

