% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoolNetPerturb.R
\name{labelState}
\alias{labelState}
\title{Label a state.}
\usage{
labelState(state, node.names, labels, rules, sep = "")
}
\arguments{
\item{state}{binary state to label}

\item{node.names}{node names of the state, the length must be the same that the state's}

\item{labels}{label that will be appended if the corresponding rule is TRUE}

\item{rules}{rules used for labelling the state. If the state satisfies the rule the corresponing label will be appended.  All the node names present in the rules must be in node.names}

\item{sep}{string to separate the labels when more than one can be applied to the state.}
}
\value{
String corresponding to the label of the state.
}
\description{
Labels a binary state using a set of labelling rules.
The elements of the state correspond the network node names. If a rule is satisfied the corresponding label is appended.
}
\examples{
state <- int2binState(0, net$genes)
labelState(state, net$genes, labels, rules, sep='')
}

