% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoolNetPerturb.R
\name{perturbPathToAttractor}
\alias{perturbPathToAttractor}
\title{Perturb network state or trajectory.}
\usage{
perturbPathToAttractor(state, net, genes, values, time = NULL,
  returnTrajectory = FALSE)
}
\arguments{
\item{state}{binary state to perturb}

\item{net}{BoolNet network to perturb}

\item{genes}{genes to perturb}

\item{values}{value of perturbed genes}

\item{time}{time of perturbation. If time=NULL the perturbation will be fixed permanently, if time=n the perturbation will last n time steps and then the rules will return to their original values.}

\item{returnTrajectory}{if TRUE returns the trajectory of state perturbation, if FALSE returns the reached attractor. Default is FALSE}
}
\value{
If returnTrajectory=TRUE returns the trajectory of the state perturbation. If FALSE returns the attractor the state reached. Al the updates are synchronous.
}
\description{
Modifies the values of a state and returns the resulting attractor or trajectory. If time=NULL the perturbation will be fixed permanently, if time=n the perturbation will last n time steps and then the rules will return to their original values.
}
\examples{
# Perturb an state by permanetly fixing the value of nodes, return the final attractor
state <- int2binState(162, cellcycle$genes)
new.attr <- perturbPathToAttractor(state, cellcycle, 
                 c('CycD', 'Rb'), c(1,0))


# Perturb an state by permanetly fixing the value of nodes, return the trajectory
state <- int2binState(162, cellcycle$genes)
new.traj <- perturbPathToAttractor(state, cellcycle, 
                 c('CycD', 'Rb'), c(1,0), 
                 returnTrajectory=TRUE)

# Perturb an state by fixing the value of nodes for one time step, return the final attractor
state <- int2binState(162, cellcycle$genes)
new.attr <- perturbPathToAttractor(state, cellcycle, 
                 c('CycD', 'Rb'), c(1,0), time=1)


# Perturb an state by fixing the value of nodes one time step, return the trajectory
state <- int2binState(162, cellcycle$genes)
new.traj <- perturbPathToAttractor(state, cellcycle, 
                 c('CycD', 'Rb'), c(1,0), time=1, 
                 returnTrajectory=TRUE)

}
\seealso{
\code{\link{...}}
}

