% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoolNetPerturb.R
\name{setStateValues}
\alias{setStateValues}
\title{Modify state, set new value to target nodes.}
\usage{
setStateValues(state, new.nodes, new.values)
}
\arguments{
\item{state}{original state}

\item{new.nodes}{nodes to modify}

\item{new.values}{new values of modified nodes}
}
\value{
State with new values in modified nodes.
}
\description{
Modify state, set new value to target nodes.
}
\examples{
state <- int2binState(0, cellcycle$genes)
new.state <- setStateValues(state, c("cycd","cyce"), c(0,1))
}

